# desafioJava.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "desafioJava",
    "description": "APIs para gerenciamento de veículos e pneus",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/api/auth/registrar": {
      "post": {
        "summary": "Registrar usuário",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              },
              "example": {
                "nome": "João Silva",
                "email": "joao@email.com",
                "senha": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Fazer login",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              },
              "example": {
                "email": "joao@email.com",
                "senha": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/veiculos": {
      "get": {
        "summary": "Listar veículos",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "Criar veículo",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              },
              "example": {
                "placa": "ABC1234",
                "marca": "Ford",
                "modelo": "Fiesta",
                "ano": 2020,
                "status": "ATIVO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/veiculos/1": {
      "get": {
        "summary": "Buscar veículo específico",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/pneus": {
      "post": {
        "summary": "Criar pneu",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              },
              "example": {
                "numeroFogo": "123456",
                "marca": "Michelin",
                "modelo": "Primacy 4",
                "medida": "205/55R16",
                "pressao": 32,
                "status": "DISPONIVEL"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/veiculos-pneus": {
      "post": {
        "summary": "Vincular pneu ao veículo",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              },
              "example": {
                "veiculoId": 1,
                "pneuId": 1,
                "posicao": "DIANTEIRA_ESQUERDA"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/veiculos-pneus/1/pneu/1": {
      "delete": {
        "summary": "Desvincular pneu do veículo",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "VeiculoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "placa": {
            "type": "string",
            "pattern": "^[A-Z]{3}[0-9][0-9A-Z][0-9]{2}$",
            "example": "ABC1D23"
          },
          "marca": {
            "type": "string",
            "example": "Volvo"
          },
          "quilometragem": {
            "type": "integer",
            "minimum": 0,
            "example": 150000
          },
          "status": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO",
              "MANUTENCAO"
            ],
            "example": "ATIVO"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VeiculoCreateDTO": {
        "type": "object",
        "required": [
          "placa",
          "marca",
          "quilometragem"
        ],
        "properties": {
          "placa": {
            "type": "string",
            "pattern": "^[A-Z]{3}[0-9][0-9A-Z][0-9]{2}$",
            "example": "ABC1D23"
          },
          "marca": {
            "type": "string",
            "example": "Volvo"
          },
          "quilometragem": {
            "type": "integer",
            "minimum": 0,
            "example": 150000
          },
          "status": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO",
              "MANUTENCAO"
            ],
            "example": "ATIVO"
          }
        }
      },
      "VeiculoComPneusDTO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/VeiculoDTO"
          },
          {
            "type": "object",
            "properties": {
              "pneusAplicados": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/VeiculoPneuDTO"
                }
              }
            }
          }
        ]
      },
      "PneuDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "numeroFogo": {
            "type": "string",
            "example": "P001"
          },
          "marca": {
            "type": "string",
            "example": "Michelin"
          },
          "pressaoAtual": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "example": 32
          },
          "status": {
            "type": "string",
            "enum": [
              "DISPONIVEL",
              "APLICADO",
              "DESGASTADO"
            ],
            "example": "DISPONIVEL"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PneuCreateDTO": {
        "type": "object",
        "required": [
          "numeroFogo",
          "marca",
          "pressaoAtual"
        ],
        "properties": {
          "numeroFogo": {
            "type": "string",
            "example": "P001"
          },
          "marca": {
            "type": "string",
            "example": "Michelin"
          },
          "pressaoAtual": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "example": 32
          },
          "status": {
            "type": "string",
            "enum": [
              "DISPONIVEL",
              "APLICADO",
              "DESGASTADO"
            ],
            "example": "DISPONIVEL"
          }
        }
      },
      "VeiculoPneuDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "veiculoId": {
            "type": "integer",
            "format": "int64"
          },
          "pneuId": {
            "type": "integer",
            "format": "int64"
          },
          "position": {
            "type": "string",
            "example": "DIANTEIRA_ESQUERDA"
          },
          "numeroFogo": {
            "type": "string",
            "example": "P001"
          },
          "marca": {
            "type": "string",
            "example": "Michelin"
          },
          "pressaoAtual": {
            "type": "number",
            "format": "double",
            "example": 32
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Erro de validação"
          },
          "message": {
            "type": "string",
            "example": "Placa já cadastrada: ABC1D23"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "properties": {}
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "desafioJava"
    }
  ],
  "security": []
}